Move more of Optics.Internal to the right places
