+++ optics-core/src/Optics/Optic.hs
@@ -4,10 +4,11 @@ module Optics.Optic
   , Is()
   , sub
   , Join()
+  , CanCompose()
   , (%)
   , (%%)
   )
   where
 
 import Optics.Internal.Optic
-import Optics.Internal.Subtyping ()
+import Optics.Internal.Subtyping (CanCompose)
